[Refactor] Refactor CCE addons update
